requirements.txt
gunicorn
meinheld
requests
pyyaml
django
djangorestframework
django-filter
markdown
whitenoise
django-material
celery
django-celery-beat
django-celery-email
drf_yasg
swagger_spec_validator
wagtail
django-guardian
django-redis
django-taggit
wechat-django[cryptography]
django-jet
django-viewflow
django-material
django-formtools
coreapi
flower
channels
